Add a fake root directory to PVR tree.
Allows user to select various parts of PVR tree in a more intuitive
manner. As screen is a fake, there is no scrollbar or screen settings
menu available on this screen.  Once user has selected a menu item, they
are presenred with the normal screen and all of the navigational items
as previously.