Optimise background image handling
fadebackground masking image has been removed and replaced with a fadediffuse animation for underlying window. Also for fullscreen windows/dialogs, underlying screen content is not shown and background image is used as the overlayed dialog/window background. Background is now not overlayed multiple times with the same content. Saving GPU resources and memory.